add fragments for RecipeStepDetail
