+++ app/build.gradle
@@ -29,5 +29,6 @@ dependencies {
     compile 'com.android.support:recyclerview-v7:25.4.0'
     compile 'com.android.support:cardview-v7:25.4.0'
     compile 'com.google.android.exoplayer:exoplayer:r2.5.0'
+    compile 'com.android.support:support-v4:25.3.1'
     testCompile 'junit:junit:4.12'
 }
