+++ app/src/main/java/com/brahmachilakala/bakingapp/RecipeStepActivity.java
@@ -1,121 +1,25 @@
 package com.brahmachilakala.bakingapp;
 
-import android.net.Uri;
 import android.os.Bundle;
-import android.support.v4.media.session.MediaSessionCompat;
-import android.support.v4.media.session.PlaybackStateCompat;
+import android.support.v4.app.FragmentTransaction;
 import android.support.v7.app.AppCompatActivity;
-import android.view.View;
-import android.widget.Button;
-import android.widget.TextView;
-
-import com.google.android.exoplayer2.ExoPlayerFactory;
-import com.google.android.exoplayer2.SimpleExoPlayer;
-import com.google.android.exoplayer2.extractor.DefaultExtractorsFactory;
-import com.google.android.exoplayer2.extractor.ExtractorsFactory;
-import com.google.android.exoplayer2.source.ExtractorMediaSource;
-import com.google.android.exoplayer2.source.MediaSource;
-import com.google.android.exoplayer2.trackselection.AdaptiveTrackSelection;
-import com.google.android.exoplayer2.trackselection.DefaultTrackSelector;
-import com.google.android.exoplayer2.trackselection.TrackSelection;
-import com.google.android.exoplayer2.trackselection.TrackSelector;
-import com.google.android.exoplayer2.ui.SimpleExoPlayerView;
-import com.google.android.exoplayer2.upstream.BandwidthMeter;
-import com.google.android.exoplayer2.upstream.DataSource;
-import com.google.android.exoplayer2.upstream.DefaultBandwidthMeter;
-import com.google.android.exoplayer2.upstream.DefaultDataSourceFactory;
-import com.google.android.exoplayer2.util.Util;
 
 import java.util.ArrayList;
 
 public class RecipeStepActivity extends AppCompatActivity {
-    private static final String TAG = RecipeStepActivity.class.getSimpleName();
-
-    private SimpleExoPlayer mExoPlayer;
-    private SimpleExoPlayerView mExoPlayerView;
-    private TextView tvStepDescription;
-    private static MediaSessionCompat mMediaSession;
-    private PlaybackStateCompat.Builder mStateBuilder;
-    private ArrayList<Step> mSteps;
-    private int position;
-    private Button mPrevBtn;
-    private Button mNextBtn;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_recipe_step);
 
-        mExoPlayerView = (SimpleExoPlayerView) findViewById(R.id.playerView);
-        mSteps = (ArrayList<Step>) getIntent().getSerializableExtra("steps");
-        position = getIntent().getIntExtra("position", 0);
+        ArrayList<Step> steps = (ArrayList<Step>) getIntent().getSerializableExtra("steps");
+        int position = getIntent().getIntExtra("position", 0);
 
-        BandwidthMeter bandwidthMeter = new DefaultBandwidthMeter();
-        TrackSelection.Factory adaptiveTrackSelectionFactory =
-                new AdaptiveTrackSelection.Factory(bandwidthMeter);
-        TrackSelector trackSelector = new DefaultTrackSelector(adaptiveTrackSelectionFactory);
-        mExoPlayer = ExoPlayerFactory.newSimpleInstance(this, trackSelector);
-        mExoPlayerView.setPlayer(mExoPlayer);
+        RecipeStepFragment fragment = RecipeStepFragment.newInstance(steps, position);
 
-        tvStepDescription = (TextView) findViewById(R.id.tv_step_description);
-
-        mPrevBtn = (Button) findViewById(R.id.bt_prev);
-        mNextBtn = (Button) findViewById(R.id.bt_next);
-
-        mPrevBtn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                position--;
-                playVideo(mSteps.get(position));
-            }
-        });
-
-        mNextBtn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                position++;
-                playVideo(mSteps.get(position));
-            }
-        });
-
-        playVideo(mSteps.get(position));
-
-
-    }
-
-    private void setupButtons() {
-        if (position <= 0) {
-            mPrevBtn.setEnabled(false);
-        }
-        if (position >= mSteps.size()) {
-            mNextBtn.setEnabled(false);
-        }
-    }
-
-    private void playVideo(Step step) {
-        setupButtons();
-        Uri videoUri = Uri.parse(step.getVideoUrl());
-        tvStepDescription.setText(step.getDescription());
-
-        DefaultBandwidthMeter defaultBandwidthMeter = new DefaultBandwidthMeter();
-        DataSource.Factory dataSourceFactory = new DefaultDataSourceFactory(this,
-                Util.getUserAgent(this, "BakingApp"), defaultBandwidthMeter);
-        ExtractorsFactory extractorsFactory = new DefaultExtractorsFactory();
-        MediaSource videoSource = new ExtractorMediaSource(videoUri,
-                dataSourceFactory, extractorsFactory, null, null);
-        mExoPlayer.prepare(videoSource);
-        mExoPlayer.setPlayWhenReady(true);
-    }
-
-    @Override
-    protected void onDestroy() {
-        super.onDestroy();
-        mExoPlayer.release();
-    }
-
-    private void releasePlayer() {
-        mExoPlayer.stop();
-        mExoPlayer.release();
-        mExoPlayer = null;
+        FragmentTransaction ft = getSupportFragmentManager().beginTransaction();
+        ft.replace(R.id.fragment_recipe_step, fragment)
+                .commit();
     }
 }
